fix(#1440): update namespace_manager import path after move to services/permissions
